Added communities

+++ CommunityManager/Controllers/MarketplaceController.cs
@@ -1,11 +1,12 @@
 ﻿using CommunityManager.Core.Contracts;
 using CommunityManager.Core.Models.Marketplace;
+using CommunityManager.Extensions;
 using CommunityManager.Infrastructure.Data.Models;
 using HouseRentingSystem.Infrastructure.Data.Common;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using System.Security.Claims;
+using static CommunityManager.Infrastructure.Data.Constants.MessageConstants;
 
 namespace CommunityManager.Controllers
 {
@@ -27,11 +28,25 @@ namespace CommunityManager.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> All()
+        public async Task<IActionResult> All(Guid id)
         {
             var model = await marketplaceService.GetAllAsync();
 
+            ViewBag.Title = "All Products";
+
             return View(model);
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> Mine()
+        {
+            var sellerId = User.Id();
+
+            var model = await marketplaceService.GetMineAsync(sellerId);
+
+            ViewBag.Title = "My Products";
+
+            return View("All", model);
         }
 
         [HttpGet]
@@ -75,6 +90,13 @@ namespace CommunityManager.Controllers
         {
             var product = await repository.GetByIdAsync<Product>(id);
 
+            if (product == null)
+            {
+                TempData[ErrorMessage] = "Incorrect product ID";
+
+                return RedirectToAction(nameof(All));
+            }
+
             ManageProductViewModel model = new ManageProductViewModel()
             {
                 Name = product.Name,
@@ -103,6 +125,13 @@ namespace CommunityManager.Controllers
         {
             var product = await repository.GetByIdAsync<Product>(id);
 
+            if (product == null)
+            {
+                TempData[ErrorMessage] = "Incorrect product ID";
+
+                return RedirectToAction(nameof(All));
+            }
+
             await marketplaceService.DeleteProductAsync(id);
 
             return RedirectToAction(nameof(All));
@@ -113,7 +142,32 @@ namespace CommunityManager.Controllers
         {
             var product = await marketplaceService.GetProductByIdAsync(id);
 
+            if (product.Name == null)
+            {
+                TempData[ErrorMessage] = "Incorrect product ID";
+
+                return RedirectToAction(nameof(All));
+            }
+
             return View(product);
+        }
+
+        public async Task<IActionResult> Buy(Guid id)
+        {
+            var product = await repository.GetByIdAsync<Product>(id);
+
+            if (product == null)
+            {
+                TempData[ErrorMessage] = "Incorrect product ID";
+
+                return RedirectToAction(nameof(All));
+            }
+
+            var buyerId = User.Id();
+
+            await marketplaceService.BuyProductAsync(id, buyerId);
+
+            return RedirectToAction(nameof(All));
         }
     }
 }
